testutil: extract test buckets into testutil
