+++ Gopkg.lock
@@ -274,6 +274,6 @@
 [solve-meta]
   analyzer-name = "dep"
   analyzer-version = 1
-  inputs-digest = "36d1abfc959a1a727e267a8fbeb3b148d6f9c081e14eb6f2f0130935af8a5ffc"
+  inputs-digest = "ccadd63894c2918adb6d95fcd37ead63d8845cd4eadfadd1c4920e771997434c"
   solver-name = "gps-cdcl"
   solver-version = 1
